+++ R/power_best_norm_rank.R
@@ -7,12 +7,9 @@
 #'
 #' This function estimates the empirical power to select the most promising group as
 #' the best group, when the outcome is normally distributed, assuming that
-#' the mean of the most promising group is at least `dif` higher than the other groups.
+#' the mean of the rest of the groups is at least `dif` lower than the most promising group.
 #' and the best group is selected by summing weighted ranks for each outcome
-#'
-#' The function allows to evaluate several outcomes at the same time, in which
-#' case evaluates if the most promising group is the best for all outcomes.
-#'
+#'#'
 #' It assume each outcome is normally distributed and they are independent between them.
 #' The outcomes may have the same mean, sd and same difference between the most promising
 #' group and the others, or they can be defined differently by outcome.